Insert integration helpers

@jock.cooper @brian.vanloo @abostani

See merge request !47

+++ spec/support/clean_db_helpers.rb
@@ -7,10 +7,10 @@ module CleanDbHelpers
     `pg_dump -O -Fc #{current_db} > #{clean_file}`
   end
 
-  def restore_clean_db(clean_file)
+  def restore_clean_db(clean_file, remove_file = true)
     self.use_transactional_fixtures = false
     `pg_restore -j 2 -O -x -c -d #{current_db} #{clean_file}`
-    `rm -f #{clean_file}`
+    `rm -f #{clean_file}` if remove_file
     ActiveRecord::Base.clear_all_connections!
     ActiveRecord::Base.reset_shared_connection
     self.use_transactional_fixtures = true
